+++ ChangeLog.md
@@ -13,7 +13,8 @@ and this project adheres to
 
 ### Added
 
-- ...
+- enhancements to `:component` module to support isolated root components 
+  that have no dependencies